refactor: move `undoableOp` to `AnkiDroid` module
* due to withCol dependency

Removes a cyclic dependency: libAnki <-> AnkiDroid

Issue 18015
