+++ AnkiDroid/src/main/java/com/ichi2/anki/BackendImporting.kt
@@ -22,10 +22,10 @@ import androidx.fragment.app.FragmentActivity
 import anki.collection.OpChangesOnly
 import anki.import_export.ImportAnkiPackageRequest
 import com.ichi2.anki.CollectionManager.withCol
+import com.ichi2.anki.observability.undoableOp
 import com.ichi2.libanki.buildSearchString
 import com.ichi2.libanki.importAnkiPackage
 import com.ichi2.libanki.importCsvRaw
-import com.ichi2.libanki.undoableOp
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
