+++ AnkiDroid/src/main/java/com/ichi2/anki/pages/PostRequestHandler.kt
@@ -28,6 +28,7 @@ import com.ichi2.anki.NoteEditor
 import com.ichi2.anki.importAnkiPackageUndoable
 import com.ichi2.anki.importCsvRaw
 import com.ichi2.anki.launchCatchingTask
+import com.ichi2.anki.observability.undoableOp
 import com.ichi2.anki.searchInBrowser
 import com.ichi2.libanki.Collection
 import com.ichi2.libanki.completeTagRaw
@@ -45,7 +46,6 @@ import com.ichi2.libanki.stats.cardStatsRaw
 import com.ichi2.libanki.stats.getGraphPreferencesRaw
 import com.ichi2.libanki.stats.graphsRaw
 import com.ichi2.libanki.stats.setGraphPreferencesRaw
-import com.ichi2.libanki.undoableOp
 import kotlinx.coroutines.Deferred
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.async
